added a confirm style page alert to ask user if they are sure they want to clear all events after clicking the button. Just as a precaution.

+++ script.js
@@ -1,18 +1,15 @@
 $(document).ready(function(){
-//  DOM variables
 
-// Javascript variables
 var hourArray = ["9AM","10AM","11AM","12PM","1PM","2PM","3PM","4PM","5PM"];
 var militaryHoursArray = ["9","10","11","12","13","14","15","16","17"];
-// function definitions
 
-
+// Displays and updates the current time live
 var currentTime = setInterval( function(){
     var time = moment().format('MMMM Do YYYY, h:mm:ss a');
     document.getElementById("currentDay").innerHTML = time;  
 });
 
-
+// Changes the clss of the time blocks depending on if they have past, are current, or are still to come.
 function updateBlocks(){
 var currentHour = moment().hours();
 $(".time-block").each(function(){
@@ -31,10 +28,14 @@ $(".time-block").each(function(){
 })
 };
 
+// Dynamically creates new rows for each time block and adds necessary information to it
 function createRows() {
     for (var i = 0; i < hourArray.length;i++) {
 
+        // gets the users input from local storage
         var savedText = localStorage.getItem(hourArray[i]);
+
+        // ensures that a block without any saved information will show it's placeholder instead of saying "null"
         if (savedText === null){
             savedText="";
         };
@@ -49,6 +50,8 @@ function createRows() {
         $(".container").append(newRow);
   
     };
+
+    // saves the users input into local storage
     $(".saveBtn").on("click", function() {
         var userInput = $(this).siblings("textarea").val();
         var hour = $(this).siblings("div").text();
@@ -57,20 +60,25 @@ function createRows() {
         localStorage.setItem(hour, userInput);
     });
      
+    // clears all saved events from local storage and clears all of the input fields
     $("#clear-events").on("click", function(event){
         event.preventDefault();
-        window.localStorage.clear();
+        // window.localStorage.clear();
+        // $(".time-block").remove();
+        // createRows();
+        var clearAll = confirm("Are you sure you would like to clear ALL of the events?");
+        if (clearAll === true){
+            window.localStorage.clear();
         $(".time-block").remove();
         createRows();
-    })
+        } else {
+            return;
+        };
+    });
 updateBlocks();
 };
 
-
-
-//  function calls
+// This will run when the page is opened
 createRows();
-// event listeners
-
 
 });
